+++ apps/gizual-app/src/primitives/canvas/canvas.module.scss.d.ts
@@ -8,5 +8,8 @@ declare const classNames: {
   readonly ControlWithLabel: "ControlWithLabel";
   readonly ControlWithLabel__Label: "ControlWithLabel__Label";
   readonly ToolbarButton: "ToolbarButton";
+  readonly MinimapContainer: "MinimapContainer";
+  readonly Minimap: "Minimap";
+  readonly DebugOverlay: "DebugOverlay";
 };
 export = classNames;