+++ apps/gizual-app/src/primitives/canvas/file-canvas/file-canvas.tsx
@@ -1,12 +1,12 @@
 import { useMainController } from "@app/controllers";
 import { Loader } from "@mantine/core";
 import { observer } from "mobx-react-lite";
-import React from "react";
+import React, { useContext } from "react";
 import { match } from "ts-pattern";
 
-import { useBlocks } from "@giz/maestro/react";
 import { FileBlock } from "../../file/block";
 import { MasonryGrid } from "../../masonry";
+import { CanvasContext } from "../canvas.context";
 import { CanvasViewModel } from "../canvas.vm";
 
 type FileCanvasProps = {
@@ -16,7 +16,7 @@ type FileCanvasProps = {
 
 export const FileCanvas = observer(({ vm, wrapper }: FileCanvasProps) => {
   const mainController = useMainController();
-  const blocks = useBlocks();
+  const blocks = useContext(CanvasContext).useBlocks();
 
   console.log({ blocks, wrapper, cw: vm.canvasWidth });
   return (